+++ plugins/woocommerce-admin/docs/features/onboarding-tasks.md
@@ -14,21 +14,21 @@ To add a custom task, you need to create a new class that extends the `Task` cla
 use Automattic\WooCommerce\Admin\Features\OnboardingTasks\Task;
 
 class MyTask extends Task {
-	public function get_id() {
-		return 'my-task';
-	}
+  public function get_id() {
+    return 'my-task';
+  }
 
-	public function get_title() {
-		return __( 'My task', 'woocommerce' );
-	}
+  public function get_title() {
+    return __( 'My task', 'woocommerce' );
+  }
 
-	public function get_content() {
-		return __( 'Add your task description here for display in the task list.', 'woocommerce');
-	}
+  public function get_content() {
+    return __( 'Add your task description here for display in the task list.', 'woocommerce');
+  }
 
-	public function get_time() {
-		return __( '2 minutes', 'woocommerce' );
-	}
+  public function get_time() {
+    return __( '2 minutes', 'woocommerce' );
+  }
 }
 ```
